chore: rename LoginPage to Login
