+++ frontend/src/router/index.js
@@ -1,7 +1,7 @@
 import Vue from "vue"
 import VueRouter from "vue-router"
 import Regist from "@/views/Regist"
-import LoginPage from "@/views/LoginPage.vue"
+import Login from "@/views/Login.vue"
 
 Vue.use(VueRouter)
 
@@ -12,8 +12,7 @@ const routes = [
   },
   {
     path: "/login",
-    name: "LoginPage",
-    component: LoginPage,
+    component: Login,
   },
 ]
